Fireaxe buff
This'll hurt if you're hit by it.

60 slash damage when wielded - hear me out.

It's rare, it's huge and can't be bagged, and it requires wielding. It swings quite slowly.

Two taps an unarmoured opponent and will three tap an armoured opponent. Doesn't benefit from Oni's +10 to blunt, as I removed the blunt damage entirely. I believe that's how that works, anyway.